Adding setup_megatron_optimizer to FabricMegatronStrategy
