+++ nemo/lightning/_strategy_lib.py
@@ -13,10 +13,11 @@
 # limitations under the License.
 
 import itertools
+import inspect
 import os
 from collections import defaultdict
 from contextlib import contextmanager
-from typing import TYPE_CHECKING, Any, Dict, Generator, Mapping, Optional, Protocol, TypeVar
+from typing import TYPE_CHECKING, Any, Dict, Generator, Mapping, Optional, Protocol, TypeVar, Callable
 
 import torch
 from torch import nn
@@ -570,3 +571,51 @@ def _sync_from_last_pipeline_stage(value: torch.Tensor, broadcast: bool = False)
                 src_rank,
                 group=parallel_state.get_pipeline_model_parallel_group(),
             )
+
+
+def setup_megatron_optimizer(
+    model, 
+    config,
+    no_weight_decay_cond: Optional[Callable] = None,
+    scale_lr_cond: Optional[Callable] = None,
+    lr_mult: float = 1.0,
+):
+    from nemo.core.optim import McoreDistributedOptimizer
+    from megatron.core.optimizer import OptimizerConfig, get_megatron_optimizer
+    
+    assert isinstance(config, OptimizerConfig), f"Expected OptimizerConfig, got {type(config)}"
+
+    class McoreOpt(McoreDistributedOptimizer):
+        def sharded_state_dict(
+            self,
+            model_sharded_state_dict,
+            optimizer_state_dict=None,
+            is_loading=False,
+            sharding_type='fully_sharded_model_space',
+        ):
+            mcore_optimizer_sig = inspect.signature(self.mcore_optimizer.sharded_state_dict).parameters
+            distrib_optim_kwargs = {}
+            if "sharding_type" in mcore_optimizer_sig:
+                distrib_optim_kwargs["sharding_type"] = sharding_type
+            state_dict = self.mcore_optimizer.sharded_state_dict(
+                model_sharded_state_dict, is_loading=is_loading, **distrib_optim_kwargs
+            )
+            return state_dict
+
+    mcore_opt = get_megatron_optimizer(
+        config,
+        list(model),
+        no_weight_decay_cond=no_weight_decay_cond,
+        scale_lr_cond=scale_lr_cond,
+        lr_mult=lr_mult,
+    )
+
+    if getattr(model.ddp_config, "overlap_param_sync", False) and getattr(
+        model.ddp_config, "align_param_gather", False
+    ):
+        param_sync_func = [model_chunk.start_param_sync for model_chunk in model]
+        param_sync_func = param_sync_func[0] if len(model) == 1 else param_sync_func
+        for module in model:
+            module.config.param_sync_func = param_sync_func
+
+    return McoreOpt(mcore_opt)
